Unfinished 1.7
This version is currently not very usable - so it's only available in the dev branch

- Added a Pluginsystem
The new pluginsystem makes it possible to reduce the amount of the basalt-user doesn't need, and makes it possible for me to add everything i want without beeing worried about the filesize of the project.
- Added a new inheritance system (Object -> VisualObject -> ChangeableObject -> actual object)
This new system makes it easily possible to inheritate from (for example) a input and modify everything you'd like - as a new object.
Changed the Frame object into multiple objects, called Container (base), BaseFrame (main frame), Frame, MovableFrame, ScrollableFrame

A lot of stuff is missing, better don't use it right now